hal/riscv64: misrify
JIRA: DO-3

+++ hal/riscv64/dtb.h
@@ -18,9 +18,9 @@
 
 #include "hal/cpu.h"
 
-#define ntoh16(x) ((((x) << 8) & 0xff00U) | (((x) >> 8) & 0xffU))
-#define ntoh32(x) ((ntoh16(x) << 16) | ntoh16((x) >> 16))
-#define ntoh64(x) ((((u64)ntoh32(x)) << 32) | (((u64)ntoh32(x)) >> 32))
+#define ntoh16(x) ((((x) << 8U) & 0xff00U) | (((x) >> 8U) & 0xffU))
+#define ntoh32(x) ((ntoh16(x) << 16U) | ntoh16((x) >> 16U))
+#define ntoh64(x) ((ntoh32(x) << 32U) | ntoh32((x) >> 32U))
 
 
 void dtb_save(void *dtb);
